+++ src/constants/DefaultFields.ts
@@ -2,5 +2,6 @@ export const DefaultFields = {
   PublishingDate: `date`,
   LastModified: `lastmod`,
   Description: `description`,
+  Title: `title`,
   Slug: `slug`
 };